+++ apps/forms/app/(www)/(home)/layout.tsx
@@ -1,0 +1,22 @@
+import { createServerComponentClient } from "@/lib/supabase/server";
+import { cookies } from "next/headers";
+import { redirect } from "next/navigation";
+
+export default async function HomeLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  const cookieStore = cookies();
+  const supabase = createServerComponentClient(cookieStore);
+
+  const { data } = await supabase.auth.getUser();
+
+  const isLoggedIn = !!data?.user;
+
+  if (isLoggedIn) {
+    redirect("/dashboard");
+  }
+
+  return <>{children}</>;
+}
